Add filtering to move area
Detect collisions to filter move area. This avoids one player to move
over another player's or player's enemy.
A method to return the collision results has also been added (this will
be used in actions in the future).Yet to be Tested

Fixed a small bug for persona.get_directions, when speed was 0.